Added Lib folder and after-30-days script
Lib folder was added; this folder was created with the intention of put
here all the functions or classes that the project could use. On adition
to that there were created different classes for a object oriented
solution. Below is the hierarchy class.

Product
└── Coverage
    ├── LowCoverage
    ├── MediumCoverage
    ├── FullCoverage
    ├── MegaCoverage
    ├── SpecialFullCoverage
    └── SuperSale

after-30-days script; this scrips was added to execute after-30-days.js
file, this run a simulation of 30 days.